Break page-metadata <-> news resolver import cycle
madge (check-circular in CI) flags the router importing the news
resolver while the news resolver imports the resolver types back. Move
the shared types to their own file.
